==> 99 Create fictional data frame.R <==

# Create fictional data frame

publications <- data.frame(author = c("AHERN, K. & DITTMAR, A. K.",
                                       "Adams and Funk",
                                       "Adams, Ferreira",
                                       "Alan et al. (2019)",
                                       "BAVOĽÁR/OROSOVÁ",
                                       "berisha et al. 2018",
                                       "2021",
                                       "carpenter et al.",
                                       "Dehejia, Wahba",
                                       "",
                                       "Jaccard et al. 2003"),
                           year = c(2012,
                                     2012,
                                     2009,
                                     2019,
                                     20015,
                                     NA,
                                     NA,
                                     NA,
                                     1999,
                                     1985,
                                     2003),
                           title = c("THE CHANGING OF THE BOARDS: THE IMPACT ON FIRM VALUATION OF MANDATED FEMALE BOARD REPRESENTATION",
                                      "Beyond the Glass Ceiling: Does Gender Matter?",
                                      "Women in the boardroom...",
                                      "understanding gender differences in leadership",
                                      "Decision-making styles...decision-making competencies...mental health",
                                      "Conv val dms measures",
                                      "Gender gaps in labor informality: The motherhood effect",
                                      "",
                                      "Causal Effects in Nonexperimental Studies: Reevaluating the Evaluation of Training Programs",
                                      "management handbook",
                                      "Interaction Effects in Multiple Regression"),
                           journal = c("THE QUARTERLY JOURNAL OF ECONOMICS",
                                        "Management Science",
                                        "J. of Financial Economics",
                                        "The Economic Journal",
                                        "Judgm. decis. mak.",
                                        "n.a.",
                                        "",
                                        "journal of management",
                                        "Journal of the American Statistical Association",
                                        "",
                                        ""),
                           vol = c(127,
                                    58,
                                    94,
                                    NA,
                                    NA,
                                    NA,
                                    NA,
                                    30,
                                    94,
                                    NA,
                                    NA),
                           issue = c(1,
                                      2,
                                      2,
                                      NA,
                                      NA,
                                      NA,
                                      NA,
                                      6,
                                      448,
                                      NA,
                                      NA),
                           publisher = c("",
                                          "n.a.",
                                          "",
                                          "",
                                          "NOT AVAILABLE",
                                          "Journal of Dynamic Decision Making",
                                          "",
                                          "",
                                          "N/A/",
                                          "",
                                          "SAGE Publications"),
                           notes = c("useful",
                                      "not useful",
                                      "need to check",
                                      "",
                                      "prueba",
                                      "encuestas",
                                      "management",
                                      "",
                                      "",
                                      "",
                                      "NA//"),
                           doi = c("article 10.1093/qje/qjr049",
                                   "DOI:10.1287/mnsc.1110.1452",
                                   "http://doi.org/10.1016/j.jfineco.2008.10.007",
                                   "https://doi.org/10.1093/ej/uez050",
                                   "indexado 10.1017/S1930297500003223",
                                   "10.11588/jddm.2018.1.43102/WOS",
                                   "WOS paper 10.1016/j.jdeveco.2020.102599",
                                   "10.1016/J.JM.2004.06.001/WOS:",
                                   "  __   10.1080/01621459.1999.10473858   *  e",
                                   "no doi available",
                                   "year-2003-10.4135/9781412984522"))
